move writeClipboardToTempFile to runtime fix eslint
