+++ src/main/deltachat/login.ts
@@ -3,8 +3,6 @@ import { getLogger } from '../../shared/logger'
 import '../notifications'
 import SplitOut from './splitout'
 import { DeltaChatAccount } from '../../shared/shared-types'
-import { stat, readdir } from 'fs/promises'
-import { join } from 'path'
 import { Context } from 'deltachat-node/node/dist/context'
 import { DesktopSettings } from '../desktop_settings'
 import { tx } from '../load-translations'
